restructure swish_crypto mod
