+++ pkm_rs/src/encryption/swish_crypto/hash.rs
@@ -1,0 +1,45 @@
+use sha1::Digest;
+use sha1::digest::OutputSizeUser;
+use sha2::Sha256;
+
+use wasm_bindgen::prelude::*;
+
+const INTRO_HASH_BYTES: [u8; 64] = [
+    0x9e, 0xc9, 0x9c, 0xd7, 0x0e, 0xd3, 0x3c, 0x44, 0xfb, 0x93, 0x03, 0xdc, 0xeb, 0x39, 0xb4, 0x2a,
+    0x19, 0x47, 0xe9, 0x63, 0x4b, 0xa2, 0x33, 0x44, 0x16, 0xbf, 0x82, 0xa2, 0xba, 0x63, 0x55, 0xb6,
+    0x3d, 0x9d, 0xf2, 0x4b, 0x5f, 0x7b, 0x6a, 0xb2, 0x62, 0x1d, 0xc2, 0x1b, 0x68, 0xe5, 0xc8, 0xb5,
+    0x3a, 0x05, 0x90, 0x00, 0xe8, 0xa8, 0x10, 0x3d, 0xe2, 0xec, 0xf0, 0x0c, 0xb2, 0xed, 0x4f, 0x6d,
+];
+
+const OUTRO_HASH_BYTES: [u8; 64] = [
+    0xd6, 0xc0, 0x1c, 0x59, 0x8b, 0xc8, 0xb8, 0xcb, 0x46, 0xe1, 0x53, 0xfc, 0x82, 0x8c, 0x75, 0x75,
+    0x13, 0xe0, 0x45, 0xdf, 0x32, 0x69, 0x3c, 0x75, 0xf0, 0x59, 0xf8, 0xd9, 0xa2, 0x5f, 0xb2, 0x17,
+    0xe0, 0x80, 0x52, 0xdb, 0xea, 0x89, 0x73, 0x99, 0x75, 0x79, 0xaf, 0xcb, 0x2e, 0x80, 0x07, 0xe6,
+    0xf1, 0x26, 0xe0, 0x03, 0x0a, 0xe6, 0x6f, 0xf6, 0x41, 0xbf, 0x7e, 0x59, 0xc2, 0xae, 0x55, 0xfd,
+];
+
+fn compute_hash(data: &[u8]) -> aes::cipher::Array<u8, <Sha256 as OutputSizeUser>::OutputSize> {
+    let with_intro_outro: Box<[u8]> = INTRO_HASH_BYTES
+        .into_iter()
+        .chain(data.iter().copied())
+        .chain(OUTRO_HASH_BYTES)
+        .collect();
+
+    Sha256::digest(&with_intro_outro)
+}
+
+#[wasm_bindgen(js_name = hashIsValid)]
+pub fn hash_is_valid(data: &[u8]) -> bool {
+    let (payload, stored_hash) = data.split_at(data.len() - super::SIZE_HASH);
+    let computed_hash = &compute_hash(payload);
+
+    stored_hash
+        .iter()
+        .zip(computed_hash)
+        .all(|(stored, computed)| *stored == *computed)
+}
+
+#[wasm_bindgen(js_name = swishCryptoHash)]
+pub fn compute_hash_js(data: &[u8]) -> Box<[u8]> {
+    compute_hash(data).iter().copied().collect()
+}
