+++ pkm_rs/src/encryption/swish_crypto/mod.rs
@@ -1,0 +1,4 @@
+mod block;
+mod hash;
+
+const SIZE_HASH: usize = 0x20;
